test: move tests into one file and add more tests
